perf: Split multicall worker (#6949)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!--
copilot:all
-->

📦🚀🔧

<!--
1.  📦 for adding a new module
2.  🚀 for adding a new feature
3.  🔧 for improving the multicall performance
-->
This pull request adds a new feature to use a worker thread for
multicall requests in the web app. This improves the performance and
reliability of fetching the best AMM trade across multiple routes.

> _`Multicall` logic_
> _Moves to `worker` thread - kire_
> _Spring of performance_

* Delegate multicall requests to worker thread using `fetchChunk`
function
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eR21),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eR61-R67),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eL68-R76),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-cb95365019f7ba2030863c54e03dafdfde7363d4004781eab13cf7e887fa31deL2-R26),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-cb95365019f7ba2030863c54e03dafdfde7363d4004781eab13cf7e887fa31deR34-R44),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-bfd062baaa17d4fb8942e2902016d461de7a5fb4c84002d53123186349e9df1eR5),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-bfd062baaa17d4fb8942e2902016d461de7a5fb4c84002d53123186349e9df1eL187-R200))
* Import `fetchChunk` function from `state/multicall/fetchChunk` module
in `hooks/useBestAMMTrade` and `quote-worker` files
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eR21),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-cb95365019f7ba2030863c54e03dafdfde7363d4004781eab13cf7e887fa31deL2-R26))
* Add generic type parameter `<T>` to `postMessage` method of
`WorkerProxy` class in `hooks/useBestAMMTrade` file to return promise of
type `T`
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eL35-R36),
[link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eL41-R42))
* Add `fetchChunk` method to `WorkerProxy` class in
`hooks/useBestAMMTrade` file to call `multicallChunk` command on worker
thread with given parameters
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eR61-R67))
* Export `worker` variable that holds an instance of `WorkerProxy` class
in `hooks/useBestAMMTrade` file
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-b799230b7acaccf91346004ce8e813994b30a79d3e04497a6ce341c7d046ec2eL68-R76))
* Add `multicallChunk` variant to `WorkerEvent` union type in
`quote-worker` file to represent command and parameters for multicall
requests
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-cb95365019f7ba2030863c54e03dafdfde7363d4004781eab13cf7e887fa31deL2-R26))
* Add branch to `onmessage` handler in `quote-worker` file to handle
`multicallChunk` command and call `fetchChunk` function with given
parameters
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-cb95365019f7ba2030863c54e03dafdfde7363d4004781eab13cf7e887fa31deR34-R44))
* Import `worker` variable from `hooks/useBestAMMTrade` module in
`state/multicall/updater` file
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-bfd062baaa17d4fb8942e2902016d461de7a5fb4c84002d53123186349e9df1eR5))
* Modify call to `fetchChunk` function in `state/multicall/updater` file
to use `fetchChunk` method of `worker` variable with chain id, chunk,
and current block as parameters
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-bfd062baaa17d4fb8942e2902016d461de7a5fb4c84002d53123186349e9df1eL187-R200))
* Define `fetchChunk` function that performs multicall requests using
`getViemClients` function in `state/multicall/fetchChunk` module
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-756d218de2219271e4b96affb244c02cc9c25021b88c7e5bc41aa07479009f3fR1-R69))
* Handle errors, retries, and gas limit splitting in `fetchChunk`
function
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-756d218de2219271e4b96affb244c02cc9c25021b88c7e5bc41aa07479009f3fR1-R69))
* Comment out original definition of `fetchChunk` function in
`state/multicall/updater` file to avoid using `useMulticallContract`
hook and `multicallContract` instance
([link](https://github.com/pancakeswap/pancake-frontend/pull/6949/files?diff=unified&w=0#diff-bfd062baaa17d4fb8942e2902016d461de7a5fb4c84002d53123186349e9df1eL16-R76))
