+++ apps/web/src/hooks/useCallWithGasPrice.ts
@@ -11,10 +11,10 @@ import {
   InferFunctionName,
   WriteContractParameters,
 } from 'viem'
+import { EstimateContractGasParameters } from 'viem/dist/types/actions/public/estimateContractGas'
 import { useWalletClient } from 'wagmi'
 import { SendTransactionResult } from 'wagmi/actions'
 import { calculateGasMargin } from 'utils'
-import { EstimateContractGasParameters } from 'viem/dist/types/actions/public/estimateContractGas'
 import { useActiveChainId } from './useActiveChainId'
 
 export function useCallWithGasPrice() {
