+++ apps/web/src/quote-worker.ts
@@ -1,10 +1,12 @@
 import { SmartRouter } from '@pancakeswap/smart-router/evm'
 import { log } from 'next-axiom'
+import { Call } from 'state/multicall/actions'
+import { fetchChunk } from 'state/multicall/fetchChunk'
 import { getViemClients } from 'utils/viem'
 
 const { parseCurrency, parseCurrencyAmount, parsePool, serializeTrade } = SmartRouter.Transformer
 
-export type WorkerEvent = [
+export type WorkerGetBestTradeEvent = [
   id: number,
   message: {
     cmd: 'getBestTrade'
@@ -41,12 +43,47 @@ const fetchWithLogging = async (url: RequestInfo | URL, init?: RequestInit) => {
 
 globalThis.fetch = fetchWithLogging
 
+export type WorkerMultiChunkEvent = [
+  id: number,
+  message: {
+    cmd: 'multicallChunk'
+    params: {
+      chainId: number
+      chunk: Call[]
+      minBlockNumber: number
+    }
+  },
+]
+
+export type WorkerEvent = WorkerGetBestTradeEvent | WorkerMultiChunkEvent
+
 const onChainQuoteProvider = SmartRouter.createQuoteProvider({ onChainProvider: getViemClients })
 
 // eslint-disable-next-line no-restricted-globals
 addEventListener('message', (event: MessageEvent<WorkerEvent>) => {
   const { data } = event
   const [id, message] = data
+  if (message.cmd === 'multicallChunk') {
+    fetchChunk(message.params.chainId, message.params.chunk, message.params.minBlockNumber)
+      .then((res) => {
+        postMessage([
+          id,
+          {
+            success: true,
+            result: res,
+          },
+        ])
+      })
+      .catch((err) => {
+        postMessage([
+          id,
+          {
+            success: false,
+            error: err,
+          },
+        ])
+      })
+  }
   if (message.cmd === 'getBestTrade') {
     const parsed = SmartRouter.APISchema.zRouterPostParams.safeParse(message.params)
     if (parsed.success === false) {
